(wip): Added some repository, models and paging implementation up to data:repository. Pending domain and ui implementation

+++ shared/data/datasources-core/src/commonMain/kotlin/dev/afalabarce/mangaref/data/datasources/core/features/characters/remote/DragonBallCharactersApi.kt
@@ -11,5 +11,5 @@ interface DragonBallCharactersApi {
     suspend fun getAllCharacters(@Query("page") page: Int, @Query("limit") limit: Int): PaginatedResult<RemoteDragonBallCharacter>
 
     @GET("/api/characters/{characterId}")
-    suspend fun getCharacter(@Path("characterId") characterId: Int): RemoteDragonBallCharacter
+    suspend fun getCharacter(@Path("characterId") characterId: Long): RemoteDragonBallCharacter
 }
